+++ src/main/java/com/example/short_kki/domain/recipe/dto/IngredientInfo.java
@@ -1,0 +1,15 @@
+package com.example.short_kki.domain.recipe.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+
+public record IngredientInfo(
+        @NotBlank(message = "재료명은 필수입니다")
+        String name,
+        @NotNull(message = "단위는 필수입니다")
+        String unit, // 단위 (g, 개, 큰술 등)
+        Integer amount
+
+) {
+
+}
